# contracts/openapi.json
{"openapi":"3.0.1","info":{"title":"Payment Service API","description":"Owns AI Credit wallets, transactions and reservations. Every payment API operation requires an authorized service identity; owner-scoped operations also require the trusted X-Payment-Owner context.","contact":{"name":"Jobseeker Copilot"},"license":{"name":"MIT"},"version":"3.1.0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/internal/system-data/seed/payments":{"post":{"tags":["payment-system-data-controller"],"operationId":"seedPayments","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemDataPaymentSeedRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemDataResult"}}}}},"security":[{"environmentDataToken":[]}]}},"/api/v1/payments/reservations":{"post":{"tags":["payment-controller"],"summary":"Reserve AI tokens before an internal LLM workflow","operationId":"createReservation","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReservationRequest"}}},"required":true},"responses":{"200":{"description":"Tokens reserved","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}},"402":{"description":"Insufficient AI Credit","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/reservations/{reservationId}/release":{"post":{"tags":["payment-controller"],"summary":"Release a reserved AI token hold","operationId":"releaseReservation","parameters":[{"name":"X-Payment-Owner","in":"header","required":true},{"name":"reservationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReleaseReservationRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReleaseReservationResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/reservations/{reservationId}/commit":{"post":{"tags":["payment-controller"],"summary":"Commit actual AI token usage for a reservation","operationId":"commitReservation","parameters":[{"name":"X-Payment-Owner","in":"header","required":true},{"name":"reservationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommitReservationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommitReservationResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/estimate":{"post":{"tags":["payment-controller"],"summary":"Estimate AI token usage for a feature","operationId":"estimate","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstimateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EstimateResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/demo-purchase":{"post":{"tags":["payment-controller"],"summary":"Demo purchase AI tokens","operationId":"demoPurchase","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemoPurchaseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DemoPurchaseResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/confirm-stripe-purchase":{"post":{"tags":["payment-controller"],"summary":"Confirm a Stripe checkout purchase and add AI tokens idempotently","operationId":"confirmStripePurchase","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmStripePurchaseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DemoPurchaseResponse"}}}}},"security":[{"serviceToken":[]}]}},"/internal/system-data/verify/payments/{userId}":{"get":{"tags":["payment-system-data-controller"],"operationId":"verifyPayments","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemDataResult"}}}}},"security":[{"environmentDataToken":[]}]}},"/internal/system-data/v1/runtime-owners/{scenarioId}/identities/{identityKey}/owners/{userId}":{"get":{"tags":["payment-system-data-controller"],"operationId":"verifyRuntimeOwner","parameters":[{"name":"scenarioId","in":"path","required":true,"schema":{"pattern":"[a-z0-9][a-z0-9-]{1,54}-v[1-9][0-9]{0,6}","type":"string"}},{"name":"identityKey","in":"path","required":true,"schema":{"pattern":"[a-z0-9][a-z0-9-]{0,54}","type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemDataResult"}}}}},"security":[{"environmentDataToken":[]}]},"delete":{"tags":["payment-system-data-controller"],"operationId":"resetRuntimeOwner","parameters":[{"name":"scenarioId","in":"path","required":true,"schema":{"pattern":"[a-z0-9][a-z0-9-]{1,54}-v[1-9][0-9]{0,6}","type":"string"}},{"name":"identityKey","in":"path","required":true,"schema":{"pattern":"[a-z0-9][a-z0-9-]{0,54}","type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemDataResult"}}}}},"security":[{"environmentDataToken":[]}]}},"/api/v1/payments/wallet":{"get":{"tags":["payment-controller"],"summary":"Get or create the user's AI token wallet","operationId":"wallet","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"responses":{"200":{"description":"Wallet summary returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WalletSummaryResponse"}}}},"401":{"description":"Service authentication failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WalletSummaryResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/transactions":{"get":{"tags":["payment-controller"],"summary":"Get the user's AI token transaction history","operationId":"transactions","parameters":[{"name":"X-Payment-Owner","in":"header","required":true},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionsResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/reservations/{reservationId}":{"get":{"tags":["payment-controller"],"summary":"Get owner-scoped reservation lifecycle and recovery evidence","operationId":"reservationStatus","parameters":[{"name":"X-Payment-Owner","in":"header","required":true},{"name":"reservationId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReservationStatusResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payments/pricing":{"get":{"tags":["payment-controller"],"summary":"Get active AI token pricing plans","operationId":"pricing","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PricingPlansResponse"}}}}},"security":[{"serviceToken":[]}]}},"/internal/system-data/scenario/{scenarioId}/payments/{userId}":{"delete":{"tags":["payment-system-data-controller"],"operationId":"resetPayments","parameters":[{"name":"scenarioId","in":"path","required":true,"schema":{"type":"string"}},{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SystemDataResult"}}}}},"security":[{"environmentDataToken":[]}]}}},"components":{"schemas":{"AiTokenReservation":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string"},"walletId":{"type":"string","format":"uuid"},"feature":{"type":"string"},"operationKey":{"type":"string"},"reservedTokens":{"type":"integer","format":"int64"},"committedTokens":{"type":"integer","format":"int64"},"releasedTokens":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["RESERVED","COMMITTED","RELEASED","FAILED"]},"referenceType":{"type":"string"},"referenceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"expiresAt":{"type":"string","format":"date-time"},"committedAt":{"type":"string","format":"date-time"},"releasedAt":{"type":"string","format":"date-time"},"lastTransitionAt":{"type":"string","format":"date-time"},"lastTransitionReason":{"type":"string"},"reconciliationAttempts":{"type":"integer","format":"int32"},"lastReconciliationAttemptAt":{"type":"string","format":"date-time"},"reconciliationErrorCode":{"type":"string"}}},"AiTokenTransaction":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"sequenceNumber":{"type":"integer","format":"int64"},"userId":{"type":"string"},"walletId":{"type":"string","format":"uuid"},"transactionType":{"type":"string","enum":["FREE_TRIAL_GRANTED","DEMO_PURCHASE","PURCHASE","SPEND","REFUND","RESERVATION","RESERVATION_RELEASED","ADJUSTMENT"]},"tokenAmount":{"type":"integer","format":"int64"},"balanceDeltaTokens":{"type":"integer","format":"int64"},"balanceBefore":{"type":"integer","format":"int64"},"balanceAfter":{"type":"integer","format":"int64"},"operationId":{"type":"string"},"description":{"type":"string"},"referenceType":{"type":"string"},"referenceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"AiTokenWallet":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string"},"balanceTokens":{"type":"integer","format":"int64"},"lifetimePurchasedTokens":{"type":"integer","format":"int64"},"lifetimeSpentTokens":{"type":"integer","format":"int64"},"lifetimeRefundedTokens":{"type":"integer","format":"int64"},"freeTrialGranted":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"SystemDataPaymentSeedRequest":{"type":"object","properties":{"scenarioId":{"type":"string"},"userId":{"type":"string"},"wallet":{"$ref":"#/components/schemas/AiTokenWallet"},"transactions":{"type":"array","items":{"$ref":"#/components/schemas/AiTokenTransaction"}},"reservations":{"type":"array","items":{"$ref":"#/components/schemas/AiTokenReservation"}}}},"SystemDataResult":{"type":"object","properties":{"service":{"type":"string"},"operation":{"type":"string"},"status":{"type":"string"},"recordsAffected":{"type":"integer","format":"int32"},"activeEnvironment":{"type":"string"},"details":{"type":"object","additionalProperties":{"type":"object"}},"warnings":{"type":"array","items":{"type":"string"}}}},"CreateReservationRequest":{"required":["feature","operationKey"],"type":"object","properties":{"feature":{"maxLength":128,"minLength":0,"type":"string"},"estimatedTokens":{"minimum":1,"type":"integer","format":"int64"},"operationKey":{"maxLength":200,"minLength":0,"pattern":"[A-Za-z0-9][A-Za-z0-9._:-]{0,199}","type":"string"},"referenceType":{"maxLength":64,"minLength":0,"type":"string"},"referenceId":{"maxLength":255,"minLength":0,"type":"string"}}},"ReservationResponse":{"type":"object","properties":{"reservationId":{"type":"string","format":"uuid"},"userId":{"type":"string"},"reservedTokens":{"type":"integer","format":"int64"},"balanceAfterReservation":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["RESERVED","COMMITTED","RELEASED","FAILED"]},"operationKey":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"}}},"ReleaseReservationRequest":{"type":"object","properties":{"reason":{"type":"string"}}},"ReleaseReservationResponse":{"type":"object","properties":{"reservationId":{"type":"string","format":"uuid"},"releasedTokens":{"type":"integer","format":"int64"},"wallet":{"$ref":"#/components/schemas/WalletSummaryResponse"}}},"WalletSummaryResponse":{"type":"object","properties":{"userId":{"type":"string"},"balanceTokens":{"type":"integer","format":"int64"},"lifetimePurchasedTokens":{"type":"integer","format":"int64"},"lifetimeSpentTokens":{"type":"integer","format":"int64"},"lifetimeRefundedTokens":{"type":"integer","format":"int64"},"freeTrialGranted":{"type":"boolean"}}},"CommitReservationRequest":{"required":["actualTokens"],"type":"object","properties":{"actualTokens":{"minimum":0,"type":"integer","format":"int64"},"provider":{"type":"string"},"model":{"type":"string"},"inputTokens":{"type":"integer","format":"int64"},"outputTokens":{"type":"integer","format":"int64"},"description":{"type":"string"}}},"CommitReservationResponse":{"type":"object","properties":{"reservationId":{"type":"string","format":"uuid"},"committedTokens":{"type":"integer","format":"int64"},"releasedTokens":{"type":"integer","format":"int64"},"wallet":{"$ref":"#/components/schemas/WalletSummaryResponse"},"spendTransaction":{"$ref":"#/components/schemas/TransactionResponse"}}},"TransactionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"transactionType":{"type":"string","enum":["FREE_TRIAL_GRANTED","DEMO_PURCHASE","PURCHASE","SPEND","REFUND","RESERVATION","RESERVATION_RELEASED","ADJUSTMENT"]},"tokenAmount":{"type":"integer","format":"int64"},"balanceDeltaTokens":{"type":"integer","format":"int64"},"balanceBefore":{"type":"integer","format":"int64"},"balanceAfter":{"type":"integer","format":"int64"},"operationId":{"type":"string"},"description":{"type":"string"},"referenceType":{"type":"string"},"referenceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"EstimateRequest":{"required":["feature"],"type":"object","properties":{"feature":{"type":"string"},"estimatedInputTokens":{"minimum":0,"type":"integer","format":"int64"},"estimatedOutputTokens":{"minimum":0,"type":"integer","format":"int64"}}},"EstimateResponse":{"type":"object","properties":{"feature":{"type":"string"},"estimatedTotalTokens":{"type":"integer","format":"int64"},"estimatedCostTokens":{"type":"integer","format":"int64"},"userBalanceTokens":{"type":"integer","format":"int64"},"canAfford":{"type":"boolean"}}},"DemoPurchaseRequest":{"required":["pricingPlanId"],"type":"object","properties":{"pricingPlanId":{"type":"string"}}},"DemoPurchaseResponse":{"type":"object","properties":{"wallet":{"$ref":"#/components/schemas/WalletSummaryResponse"},"transaction":{"$ref":"#/components/schemas/TransactionResponse"}}},"ConfirmStripePurchaseRequest":{"required":["pricingPlanId","stripeSessionId","userId"],"type":"object","properties":{"userId":{"type":"string"},"pricingPlanId":{"type":"string"},"tokenAmount":{"minimum":1,"type":"integer","format":"int64"},"stripeSessionId":{"type":"string"},"stripePaymentIntentId":{"type":"string"}}},"TransactionsResponse":{"type":"object","properties":{"userId":{"type":"string"},"transactions":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}},"ReservationStatusResponse":{"type":"object","properties":{"reservationId":{"type":"string","format":"uuid"},"userId":{"type":"string"},"operationKey":{"type":"string"},"feature":{"type":"string"},"reservedTokens":{"type":"integer","format":"int64"},"committedTokens":{"type":"integer","format":"int64"},"releasedTokens":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["RESERVED","COMMITTED","RELEASED","FAILED"]},"referenceType":{"type":"string"},"referenceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"expiresAt":{"type":"string","format":"date-time"},"committedAt":{"type":"string","format":"date-time"},"releasedAt":{"type":"string","format":"date-time"},"lastTransitionAt":{"type":"string","format":"date-time"},"lastTransitionReason":{"type":"string"},"reconciliationAttempts":{"type":"integer","format":"int32"},"lastReconciliationAttemptAt":{"type":"string","format":"date-time"},"reconciliationErrorCode":{"type":"string"}}},"PricingPlansResponse":{"type":"object","properties":{"plans":{"type":"array","items":{"$ref":"#/components/schemas/TokenPricingPlanResponse"}}}},"TokenPricingPlanResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"tokenAmount":{"type":"integer","format":"int64"},"priceGbpPence":{"type":"integer","format":"int64"}}}},"securitySchemes":{"serviceToken":{"type":"apiKey","description":"Dedicated service identity; caller permissions are route-scoped.","name":"X-Service-Token","in":"header"},"environmentDataToken":{"type":"apiKey","description":"Dedicated credential for isolated environment-data operations.","name":"X-Environment-Data-Token","in":"header"}}}}